add usage
git-svn-id: https://plowshare.googlecode.com/svn/trunk@751 51fab9ec-ea3e-11dd-80ed-bda8ba2650cd

+++ examples/plowdown_parallel.sh
@@ -58,11 +58,12 @@ get_modules() {
 
 # Main
 
+test $# -ge 1 || { debug "Usage: $(basename $0) FILEWITHLINKS"; exit 1; }
 trap "kill 0" SIGINT
 INFILE=$1
 PIDS=()
 while read MODULE URLS; do
-  debug "Run: plowdown $URLS"
+  debug "Downloading: $URLS"
   plowdown $URLS &
   PIDS=(${PIDS[*]} $!)
 done < <(get_modules "$INFILE")
